% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticplots.R
\name{behaviour_plot}
\alias{behaviour_plot}
\title{Output Plotting}
\usage{
behaviour_plot(ems, points)
}
\arguments{
\item{ems}{A set of \code{\link{Emulator}} objects.}

\item{points}{A set of points at which to evaluate the emulator expectation}
}
\value{
The dependency plots.
}
\description{
A simple diagnostic plot to determine relationships between emulated outputs and
each individual input dimension. If \code{points} is not provided, then a sample
of points across the full space is created by uniform sampling.
}
\examples{
 behaviour_plot(SIREmulators$ems, SIRSample$training)
 behaviour_plot(SIREmulators$ems$nS)
}
