% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{classification_diag}
\alias{classification_diag}
\title{Classification Diagnostic}
\usage{
classification_diag(
  emulator,
  targets,
  validation = NULL,
  cutoff = 3,
  plt = T,
  ...
)
}
\arguments{
\item{emulator}{An \code{\link{Emulator}} object.}

\item{targets}{The targets to compare suitability against.}

\item{validation}{The validation data.frame containing all inputs and outputs.}

\item{cutoff}{The cutoff for the implausibility measures.}

\item{plt}{Should the plot be produced?}

\item{...}{Dummy parameters for compatibility with the diagnostic wrapper.}
}
\value{
A data.frame of points that fail the diagnostic check.
}
\description{
Checks for emulator misclassifications.
}
\details{
Both the emulator implausibility and the simulator implausibility can be computed
and compared. Points for which the emulator implausibility is outside the desired cutoff
but for which the simulator implausibility is not are misclassified points.

If \code{plt = TRUE}, then the emulator and simulator implausibilities are plotted
against one another. Misclassified points (i.e. those in the bottom right 'quadrant')
are highlighted in red.
}
\examples{
classification_diag(sample_emulators$ems$nS, sample_emulators$targets, GillespieValidation)
## An empty data.frame.
classification_diag(sample_emulators$ems$nS, sample_emulators$targets, GillespieValidation,
 cutoff = 1)
## An empty data.frame.
}
