% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelatorDocumentation.R
\name{Correlator}
\alias{Correlator}
\title{Correlation Structure}
\description{
Creates a correlation structure, with the necessary specifications.

    The correlator has three main elements: the type of correlator, the associated
    hyperparameters, and the nugget term. The nugget term is broadly separate from
    the other two parameters, being type-independent.
}
\section{Constructor}{
 \code{Correlator$new(corr, hp, nug)}
}

\section{Arguments}{


    \code{corr} The type of correlation function. This is provided as a string which
    corresponds exactly with a function - the function should take three arguments
    \code{x, xp, hp}. This gives a correlation function u(x, xp) defined by
    hyperparameters \code{hp}. For a simple example, see \code{\link{exp_sq}}.

    \code{hp} The associated hyperparameters needed to define the correlation
    structure, as a named list. In the case of \code{exp_sq}, this is a list of
    one element, \code{list(theta)}.

    \code{nug} The size of the nugget term. In situations where not all variables
    are active, the main part of u(x) operates only on the active parts, xA. The
    presence of the nugget term accounts for the fact that points at the same
    position in the active space need not be at the same position in the full space.

    By default, \code{Correlator$new()} initalises with \code{corr = exp_sq},
    \code{hp = list(theta = 0.1)}, and \code{nug = 0}.
}

\section{Accessor Methods}{


    \code{get_corr(x, xp = NULL, actives = TRUE)} Returns the correlation
    between two points. If \code{xp} is \code{NULL}, then this is correlation
    between a point and itself (i.e. 1). All variables are assumed to be active
    unless otherwise stated in \code{actives}.

    \code{get_hyper_p()} Returns the list of hyperparameters.

    \code{print()} Produces a summary of the correlation structure specification.
}

\section{Object Methods}{


    \code{set_hyper_p(hp, nugget)} Modifies the hyperparameter and/or nugget
    terms. Returns a new \code{Correlator} object.
}

\examples{
test_corr <- Correlator$new(nug = 0.1)
test_corr
point1 <- c(a = 0.1, b = 0.2, c = 0.3)
point2 <- c(a = 0.15, b = 0.18, c = 0.295)
test_corr$get_corr(point1) #> 1
test_corr$get_corr(point1, point2) #> 0.6717557
test_corr$get_corr(point1, point2, actives = c(TRUE, TRUE, FALSE)) #> 0.6734372

new_corr <- test_corr$set_hyper_p(list(theta = 0.5), nug = 0.01)
new_corr$get_corr(point1, point2) #> 0.9784845
new_corr$get_corr(point1, point2, actives = c(TRUE, TRUE, FALSE)) #> 0.9785824
}
