% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostic_pass}
\alias{diagnostic_pass}
\title{Automated Diagnostics and Modifications}
\usage{
diagnostic_pass(ems, targets, validation, verbose = interactive())
}
\arguments{
\item{ems}{The emulators to consider}

\item{targets}{The output targets to compare implausibility against}

\item{validation}{The set of validation points (either a single data.frame or one per emulator)}

\item{verbose}{Whether messages should be printed while running}
}
\value{
A collection of modified emulators, potentially a subset of the original collection
}
\description{
Perform a set of diagnostics on emulators, changing them if needed.
}
\details{
There are a number of different characteristics that emulators might possess
that give rise to diagnostic flags. This function collects together some of those
whose resulting modifications can be automated. The tests, and consequences, are
as follows.

\describe{
 \item{Structured Input Space}{Looks for errors with dependence on input parameters. If
  found, the emulator's correlation length is reduced (to a minimum of 1/3);}
 \item{Structured Output Space}{Looks for errors with dependence on output value. If
  found, the training and validation data is resampled and emulators are retrained,
  to try to incorporate/remove high leverage points;}
 \item{Misclassification}{Checks agreement between emulator and simulator implausibility
  classification. If they do not match, emulator uncertainty is inflated;}
 \item{Comparison}{Checks that the emulator predictions agree with the simulator
  predictions at the validation points, allowing for expected margin of error.}
}

If the automated modifications are not sufficient to remove problems, then
offending emulators are removed from the set under consideration. Emulators in
this category should be carefully considered and their outputs analyzed: they may
require manual determination of the regression surface or additional training
points in the neightbourhood of the problematic inputs.
}
\examples{
 new_ems <- diagnostic_pass(SIREmulators$ems, SIREmulators$targets, SIRSample$validation)
}
