% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticplots.R
\name{effect_strength}
\alias{effect_strength}
\title{Find Effect Strength of Active Variables}
\usage{
effect_strength(ems, line_plot = FALSE, labels = TRUE, quadratic = TRUE)
}
\arguments{
\item{ems}{The Emulator object(s) to be analyzed.}

\item{line_plot}{Should a line plot be produced?}

\item{labels}{Whether or not the legend should be included.}

\item{quadratic}{Whether or not quadratic effect strength should be calculated.}
}
\value{
A list of data.frames: the first is the linear strength, and the second quadratic.
}
\description{
Collates the linear and quadratic contributions of the active variables to the global
emulators' behaviour
}
\details{
For a set of emulators, it can be useful to see the relative contributions of various
parameters to the global part of the emulator (i.e. the regression surface). This
function extracts the relevant information from a list of emulator objects.

The parameter \code{quadratic} controls whether quadratic effect strength is
calculated and plotted (an unecessary plot if, say, linear emulators have been trained).
The remaining options control visual aspects of the plots: \code{line_plot} determines
whether a line or bar (default) plot should be produced, and \code{labels} determines
if a legend should be provided with the plot (for large numbers of emulators, it is
advisable to set this to \code{FALSE}).
}
\examples{
 effect <- effect_strength(sample_emulators$ems)
 effect_line <- effect_strength(sample_emulators$ems, line_plot = TRUE)
}
