% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{exp_sq}
\alias{exp_sq}
\title{Exponential squared correlation function}
\usage{
exp_sq(x, xp, hp)
}
\arguments{
\item{x}{A numeric position vector}

\item{xp}{A numeric position vector}

\item{hp}{The hyperparameter theta (correlation length)}
}
\value{
The exponential-squared correlation between x and xp.
}
\description{
For points \code{x}, \code{xp} and a correlation length \code{theta}, gives the exponent
of the squared distance between \code{x} and \code{xp}, weighted by \code{theta} squared.
}
\examples{
exp_sq(1,2, list(theta = 0.1))
#> 3.720076e-44
exp_sq(c(1,2,-1),c(1.5,2.9,-0.7), list(theta = 0.2))
#> 3.266131e-13
}
