% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GillespieImplausibility}
\alias{GillespieImplausibility}
\title{Sample Implausibility Data}
\format{
A data frame with 1000 rows and 7 variables:
\describe{
  \item{aSI}{Infection: transition rate from S to I}
  \item{aIR}{Recovery: transition rate from I to R}
  \item{aSR}{Immunisation: transition rate from S to R}
  \item{nS}{Implausibility for nS}
  \item{nI}{Implausibility for nI}
  \item{nR}{Implausibility for nR}
  \item{I}{Maximum implausibility}
}
}
\usage{
GillespieImplausibility
}
\description{
A dataset containing 1000 points from the region bounded by
[0.1, 0.8], [0, 0.5], [0, 0.05] for aSI, aIR and aSR respectively.
Implausibility has been calculated (for emulators trained on the
\code{\link{GillespieSIR}} dataset) for each of the outputs nS, nI, nR,
and the maximum implausibility is included.
The target values used in calculating implausibility were:
nS: 281 (sigma 10.43);
nI: 30 (sigma 11.16);
nR: 689 (sigma 14.32)
}
\keyword{datasets}
