% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GillespieMultiWaveEmulators}
\alias{GillespieMultiWaveEmulators}
\title{Sample Multi-wave Emulators}
\format{
A list containing \code{\link{Emulator}} objects:
\describe{
  \item{Wave 1}{Emulators trained on GillespieSIR to generate wave 2 points}
  \item{Wave 2}{Emulators trained on the results of the above wave 2 points}
  \item{Wave 3}{Emulators trained on the results of the wave 3 points}
}
}
\usage{
GillespieMultiWaveEmulators
}
\description{
An rda object containing three waves of emulation on the
Gillespie SIR model.
}
\keyword{datasets}
