% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GillespieSIR}
\alias{GillespieSIR}
\title{Sample SIR data}
\format{
A data frame with 30 rows and 6 variables:
\describe{
  \item{aSI}{Infection: transition rate from S to I}
  \item{aIR}{Recovery: transition rate from I to R}
  \item{aSR}{Immunisation: transition rate from S to R}
  \item{nS}{Final number of S}
  \item{nI}{Final number of I}
  \item{nR}{Final number of R}
}
}
\usage{
GillespieSIR
}
\description{
A small dataset containing points generated using the Gillespie algorithm.
The SIR model contains three input parameters, and generates three output
parameters. The initial populations are 950 susceptible (S), 50 infected (I),
and 0 recovered (R). The final values are taken at time t=20.
}
\keyword{datasets}
