% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GillespieValidation}
\alias{GillespieValidation}
\title{Sample SIR validation data}
\format{
A data frame with 60 rows and 6 variables:
\describe{
  \item{aSI}{Infection: transition rate from S to I}
  \item{aIR}{Recovery: transition rate from I to R}
  \item{aSR}{Immunisation: transition rate from S to R}
  \item{nS}{Final number of S}
  \item{nI}{Final number of I}
  \item{nR}{Final number of R}
}
}
\usage{
GillespieValidation
}
\description{
A small dataset containing points generated using the Gillespie algorithm.
Very similar to \code{\link{GillespieSIR}}, slightly larger in size.
}
\keyword{datasets}
