% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{matern}
\alias{matern}
\title{Matern correlation function}
\usage{
matern(x, xp, hp)
}
\arguments{
\item{x}{A numeric position vector}

\item{xp}{A numeric position vector}

\item{hp}{The hyperparameters nu (smoothness) and rho (correlation length), as a named list}
}
\value{
The Matern correlation between x and xp.
}
\description{
For points \code{x}, \code{xp}, and a pair of hyperparameters \code{nu} and \code{rho}, gives
the Matern correlation between the two points.
}
\details{
At present, only half-integer arguments for nu are supported.
}
\examples{
matern(1, 2, list(nu = 1.5, rho = 0.1))
#> 5.504735e-07
matern(c(1,2,-1), c(1.5, 2.9,-0.7), list(nu = 1.5, rho = 0.2))
#> 0.0009527116
}
