% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implausibility.R
\name{nth_implausible}
\alias{nth_implausible}
\title{nth Maximum Implausibility}
\usage{
nth_implausible(ems, x, z, n = 1, max_imp = 20, cutoff = NULL)
}
\arguments{
\item{ems}{A set of \code{\link{Emulator}} objects.}

\item{x}{An input point, or \code{data.frame} of points.}

\item{z}{The target values (one for each output)}

\item{n}{The implausibility level to return.}

\item{max_imp}{A maximum implausibility to return (often used with plotting)}

\item{cutoff}{A numeric value, or vector of such, representing allowed implausibility}
}
\value{
Either the nth maximum implausibilities, or booleans (if cutoff is given).
}
\description{
For a collection of emulators, it can be helpful to combine the implausibility
measures for a given set of observations. The maximum implausibility of a point,
given a set of univariate emulators and an associated collection of target values,
is the largest implausibility of the collected set of implasusibilities. The 2nd
maximum is the maximum of the set without the largest value, and so on.
}
\examples{
# A single point
nth_implausible(sample_emulators$ems, data.frame(aSI = 0.4, aIR = 0.25, aSR = 0.025),
 sample_emulators$targets)
# A data.frame of points
grid <- expand.grid(
 aSI = seq(0.1, 0.8, length.out = 4),
 aIR = seq(0, 0.5, length.out = 4),
 aSR = seq(0, 0.05, length.out = 4)
)
# Vector of numerics
i1 <- nth_implausible(sample_emulators$ems, grid, sample_emulators$targets)
# Vector of booleans (same as i1 <= 3)
i2 <- nth_implausible(sample_emulators$ems, grid, sample_emulators$targets, cutoff = 3)
# Throws a warning as n > no. of targets
i3 <- nth_implausible(sample_emulators$ems, grid, sample_emulators$targets, n = 4)
# Vector of booleans (note different output to i2)
i4 <- nth_implausible(sample_emulators$ems, grid, sample_emulators$targets,
 cutoff = c(4, 2.5, 1))
}
