% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_emulators}
\alias{sample_emulators}
\title{Sample Emulators}
\format{
A list containing two objects:
\describe{
 \item{ems}{The trained \code{\link{Emulator}} objects.}
 \item{targets}{The targets to match to, as a set of val-sigma pairs.}
}
}
\usage{
sample_emulators
}
\description{
An RData object containing three trained emulators, and the associated
targets, for the Gillespie SIR example. The emulators have been trained
on the \code{\link{GillespieSIR}} dataset using methods documented in
this package.
}
\keyword{datasets}
