% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{summary_diagnostic}
\alias{summary_diagnostic}
\title{Summary Statistics for Emulators}
\usage{
summary_diagnostic(em, validation)
}
\arguments{
\item{em}{The emulator to test}

\item{validation}{The validation set, consisting of points and output(s)}
}
\value{
Whether the observed value lies within 3-sigma of the expected value..
}
\description{
Generates measures for emulator quality
}
\details{
A couple of summary statistics can be generated for emulators, based on their
prediction errors on a validation set. This function produces the test statistic
for a comparison to a relevant chi-squared distribution, and the similar test
statistic for an F-distribution. In both cases, the expectation and standard
deviation of the underlying distribution are also provided.

The output of this function is a logical vector stating whether the derived
value lies within 3-sigma of the expected value. In systems where errors are
expected to be correlated, higher weight should be given to the Mahalanobis
measure than the chi-squared measure. Any anomalous results can be investigated
in more depth using the \code{\link{individual_errors}} function.
}
\examples{
 summary_diagnostic(sample_emulators$ems$nR, GillespieValidation)
}
