% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{validation_diagnostics}
\alias{validation_diagnostics}
\title{Emulator Diagnostics}
\usage{
validation_diagnostics(
  emulators,
  validation,
  targets = NULL,
  which_diag = "all",
  ...
)
}
\arguments{
\item{emulators}{A list of \code{\link{Emulator}} objects.}

\item{validation}{The validation set, containing all inputs and outputs.}

\item{targets}{The list of observations for the outputs}

\item{which_diag}{Which diagnostics should be performed (see description)}

\item{...}{Any additional parameters to pass to the diagnostics (eg sd, cutoff, ...)}
}
\value{
A data.frame containing points that failed one or more diagnostic tests.
}
\description{
Performs the standard set of validation diagnostics on emulators.
}
\details{
All the diagnostics here assume the existence of a validation (or 'hold-out') dataset
from the simulator. The presence of a set of targets is optional for some checks but
mandatory for others: the appropriate warnings will be given in the event that some
checks cannot be applied.

The current options for diagnostics (with the codes for \code{which_diag}) are:

  Standardised Errors (se)

  Comparison Diagnostics (cd)

  Classification Errors (ce)

  All of the above (all)

For details of each of the tests, see the help files for \code{\link{standard_errors}},
\code{\link{comparison_diag}} and \code{\link{classification_diag}} respectively.
}
\examples{
validation_diagnostics(sample_emulators$ems, GillespieValidation, sample_emulators$targets)
# data.frame of failed points and a 3x3 set of plots
validation_diagnostics(sample_emulators$ems, GillespieValidation, sample_emulators$targets,
 c('ce','cd'))
# data.frame and a 3x2 set of plots
validation_diagnostics(sample_emulators$ems, GillespieValidation, sample_emulators$targets,
 cutoff = 2, sd = 2)
}
