% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticplots.R
\name{validation_pairs}
\alias{validation_pairs}
\title{Validation Set Diagnostics and Implausibility}
\usage{
validation_pairs(ems, points, targets, ranges, cb = FALSE)
}
\arguments{
\item{ems}{The \code{\link{Emulator}} object(s).}

\item{points}{The set of validation points to plot.}

\item{targets}{The set of targets to match to.}

\item{ranges}{If provided, this gives the plotting region (see above).}

\item{cb}{Whether or not the colour scheme should be colourblind friendly.}
}
\value{
A ggplot object.
}
\description{
Creates pairs plots on the set of validation points of diagnostic suitability and
implausibility.
}
\details{
The plots are organised as follows:

a) Emulated versus simulated output (lower diagonal). This is similar in spirit to
\code{\link{comparison_diag}}: the plotted points are their location in the input
space and the points are coloured by the emulator prediction's deviation from the
simulator value.

b) Implausibility (upper diagonal). The points are again plotted based on their
location in input space, but their colouration is now based on the implausibility
of the point.

If \code{ranges} is provided, then the plotting region is created relative to these
ranges. This can be useful if on later waves of a history match and the plotting is
to be done relative to the original input space, rather than the (reduced) parameter
space upon which the emulators have been trained.
}
\examples{
 validation_pairs(sample_emulators$ems, GillespieValidation, sample_emulators$targets)
 wider_ranges <- purrr::map(sample_emulators$ems[[1]]$ranges, ~.*c(-2, 2))
 validation_pairs(sample_emulators$ems$nS, GillespieValidation,
  sample_emulators$targets, ranges = wider_ranges, cb = TRUE)
}
