% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelbuilding.R
\name{variance_emulator_from_data}
\alias{variance_emulator_from_data}
\title{Variance Emulator Creation}
\usage{
variance_emulator_from_data(
  input_data,
  output_names,
  ranges,
  input_names = names(ranges),
  bimodal_cut = NULL,
  bimodal_region = NULL,
  ...
)
}
\arguments{
\item{input_data}{All model runs at all points.}

\item{output_names}{The observation names.}

\item{ranges}{A named list of parameter ranges}

\item{input_names}{The names of the parameters (if \code{ranges} is not provided).}

\item{bimodal_cut}{The cutoff values for bimodality filtering}

\item{bimodal_region}{Whether to keep points above (1) or below (-1) the cuts.}

\item{...}{Optional parameters that can be passed to \code{link{emulator_from_data}}.}
}
\value{
A list of lists: one for the variance emulators and one for the function emulators.
}
\description{
For stochastic systems, it can be helpful to emulate the variance as well as the function.
This is particularly true if one expects the variance to be very different in different
areas of the parameter space (for example, in an epidemic model). This function performs
the requisite two-stage Bayes Linear update.
}
\details{
All observations are required (including replicates at points) - this function collects
them into the required chunks and calculates the summary statistics as required. If bimodality
is suspected in the data, and in particular in a situation where the bimodality is one
of epidemic die out vs explosion, cutoffs can be supplied for each output: two sets of
parameters should be passed. The \code{bimodal_cut} gives a feasible boundary defining the
region of interest for each output; the \code{bimodal_region} determines whether the
region of interest is above the cutoff point (1) or below the cutoff point (-1) for
each output. These parameters should be passed as numeric vectors, and must have the
same length as \code{output_names}.

All other parameters passed to this function are equivalent to those in
\code{\link{emulator_from_data}} - one notable absence is that by default, the returned
emulators are the Bayes Linear adjusted forms.
}
\examples{
 # A simple example using the BirthDeath dataset
 v_ems <- variance_emulator_from_data(BirthDeath$training, c("Y"),
  list(lambda = c(0, 0.08), mu = c(0.04, 0.13)), c_lengths = c(0.75))

}
