% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelbuilding.R
\name{variance_emulator_from_data_old}
\alias{variance_emulator_from_data_old}
\title{Variance Emulator Creation (Deprecated)}
\usage{
variance_emulator_from_data_old(
  input_data_var,
  input_data_exp,
  npoints,
  output_names,
  ranges,
  input_names = names(ranges),
  kurt = 3,
  beta,
  u,
  c_lengths,
  funcs,
  deltas,
  ev,
  quadratic = TRUE,
  beta.var = FALSE
)
}
\arguments{
\item{input_data_var}{The variance data.}

\item{input_data_exp}{The function data.}

\item{npoints}{The number of replicates per observation.}

\item{output_names}{The observation names.}

\item{ranges}{A named list of parameter ranges}

\item{input_names}{The names of the parameters (if \code{ranges} is not provided).}

\item{kurt}{The expected kurtosis of the data.}

\item{beta}{A list of regression coefficients for each output.}

\item{u}{A list of \code{\link{Correlator}} objects for each output.}

\item{c_lengths}{A list of correlation lengths for each output.}

\item{funcs}{A list of regression functions for each output.}

\item{deltas}{Nugget terms for each correlation structure.}

\item{ev}{Estimates of ensemble variability for each output.}

\item{quadratic}{Should a quadratic or linear fit be found?}

\item{beta.var}{Should regression coefficient uncertainty be included?}
}
\value{
A list of lists: one for the variance emulators and one for the function emulators.
}
\description{
For stochastic systems, it can be helpful to emulate the variance as well as the function.
This is particularly true if one expects the variance to be very different in different
areas of the parameter space (for example, in an epidemic model). This function performs
the requisite two-stage Bayes Linear update.
}
\details{
Two sets of data are required: the observed variance of the stochastic runs at each point,
and their corresponding mean output values. The data.frames should have the same named
inputs and outputs. A parameter \code{npoints} is also required: this is a single numeric
or vector of numerics representing the number of replicates done at each input point.

All other parameters passed to this function are equivalent to those in
\code{\link{emulator_from_data}} - one notable absence is that by default, the returned
emulators are the Bayes Linear adjusted forms.
}
\examples{
 # Use the BirthDeath dataset
 ranges <- list(lambda = c(0, 0.08), mu = c(0.04, 0.13))
 v_ems <- variance_emulator_from_data_old(BirthDeath$var, BirthDeath$mean, BirthDeath$reps,
  paste0('t', c(1, 7, 15)), ranges)
 v_ems$variance
 v_ems$expectation
}
