% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiwaveplot.R
\name{wave_variance}
\alias{wave_variance}
\title{Emulator Variance across waves}
\usage{
wave_variance(
  waves,
  output_names,
  plot_dirs = names(waves[[1]][[1]]$ranges)[1:2],
  wave_numbers = 1:length(waves),
  ppd = 20,
  sd = FALSE
)
}
\arguments{
\item{waves}{A list of lists of \code{\link{Emulator}} objects, corresponding to the waves}

\item{output_names}{The list of desired outputs to be plotted}

\item{plot_dirs}{The (two) input parameters to be plotted.}

\item{wave_numbers}{A numeric vector of which waves to plot.}

\item{ppd}{The number of grid points per plotting dimension.}

\item{sd}{Should the standard deviation be plotted instead of the variance? Default: FALSE}
}
\value{
The ggplot object(s).
}
\description{
Plots the emulator variance for each output across emulator waves.
}
\details{
It is instructive to look at the change in emulator variance over successive
waves, rather than across successive outputs. This function provides a means
of doing so quickly for each emulator output.

A 2d slice is taken across the input space, where mid-range values of any
non-plotted parameters are fixed. The emulator variance (or standard
deviation) is then calculated across the two-dimensional subspace for each
wave, and for each output.
}
\examples{
 outputs <- c('nS', 'nI', 'nR')
 wave_variance(SIRMultiWaveEmulators, outputs, ppd = 5)
 wave_variance(SIRMultiWaveEmulators, c('nI', 'nR'),
  plot_dirs = c('aIR', 'aSR'), ppd = 5, sd = TRUE)

}
