% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{comparison_diag}
\alias{comparison_diag}
\title{Comparison Diagnostics}
\usage{
comparison_diag(emulator, targets, validation, sd = 3, plt = interactive())
}
\arguments{
\item{emulator}{The emulator in question}

\item{targets}{The output targets}

\item{validation}{The validation set}

\item{sd}{The range of uncertainty allowed}

\item{plt}{Whether to plot or not}
}
\value{
A data.frame of failed points
}
\description{
Shorthand function for diagnostic test `cd'.
}
\details{
For details of the function, see \code{\link{get_diagnostic}} and for the plot
see \code{\link{analyze_diagnostic}}.
}
\references{
Jackson (2018) <http://etheses.dur.ac.uk/12826>
}
\seealso{
\code{\link{get_diagnostic}}, \code{\link{analyze_diagnostic}},
 \code{\link{validation_diagnostics}}

Other diagnostic functions: 
\code{\link{analyze_diagnostic}()},
\code{\link{classification_diag}()},
\code{\link{get_diagnostic}()},
\code{\link{individual_errors}()},
\code{\link{residual_diag}()},
\code{\link{standard_errors}()},
\code{\link{summary_diag}()},
\code{\link{validation_diagnostics}()}
}
\concept{diagnostic functions}
