% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{standard_errors}
\alias{standard_errors}
\title{Standardized Error Diagnostics}
\usage{
standard_errors(
  emulator,
  targets = NULL,
  validation = NULL,
  plt = interactive()
)
}
\arguments{
\item{emulator}{The emulator in question}

\item{targets}{The output targets}

\item{validation}{The validation set}

\item{plt}{Whether to plot or not}
}
\value{
A data.frame of failed points
}
\description{
Shorthand function for diagnostic test `se'.
}
\details{
For details of the function, see \code{\link{get_diagnostic}} and for the plot
see \code{\link{analyze_diagnostic}}.
}
\references{
Jackson (2018) <http://etheses.dur.ac.uk/12826>
}
\seealso{
\code{\link{get_diagnostic}}, \code{\link{analyze_diagnostic}},
 \code{\link{validation_diagnostics}}

Other diagnostic functions: 
\code{\link{analyze_diagnostic}()},
\code{\link{classification_diag}()},
\code{\link{comparison_diag}()},
\code{\link{get_diagnostic}()},
\code{\link{individual_errors}()},
\code{\link{residual_diag}()},
\code{\link{summary_diag}()},
\code{\link{validation_diagnostics}()}
}
\concept{diagnostic functions}
