% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiwaveplot.R
\name{wave_values}
\alias{wave_values}
\title{Multiple Wave Output Plotting}
\usage{
wave_values(
  waves,
  targets,
  output_names = names(targets),
  surround = FALSE,
  restrict = FALSE,
  p_size = 1.5,
  l_wid = 1.5
)
}
\arguments{
\item{waves}{The list of data.frames, one for each set of outputs at that wave.}

\item{targets}{The output targets.}

\item{output_names}{The outputs to plot.}

\item{surround}{As in \code{\link{wave_points}}.}

\item{restrict}{Should the plotting automatically restrict to failing target windows?}

\item{p_size}{As in \code{\link{wave_points}}.}

\item{l_wid}{The width of the lines that create the target boxes.}
}
\value{
A ggplot object.
}
\description{
Given multiple waves of points, produces pairs plots of the outputs.
}
\details{
This function operates in a similar fashion to \code{\link{wave_points}} - the main
difference is that the output values are plotted. Consequently, the set of targets is required
to overlay the region of interest onto the plot.
}
\examples{
 wave_values(GillespieMultiWaveData, sample_emulators$targets, surround = TRUE, p_size = 1)
 wave_values(GillespieMultiWaveData, sample_emulators$targets, c('nS', 'nI'), l_wid = 0.8)
}
